Bug 1000935: Added SNEP PUT sending and receiving
SNEP PUT is the SNEP request for sending NDEF messages to an NFC
peer. This commit adds functions for sending and receiving PUT
request. The send function sends an NDEF message to the peer, as
specified by SNEP. The receive function returns the data that
was transfered with the last SNEP PUT that was received from the
peer.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>